api_comentarios/src/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
requests
pytest 
